Refactor FrontPage
