+++ Sources/LightPortal/front/PageArticle.php
@@ -1,0 +1,172 @@
+<?php
+
+namespace Bugo\LightPortal\Front;
+
+use Bugo\LightPortal\Helpers;
+use Bugo\LightPortal\Subs;
+
+/**
+ * PageArticle.php
+ *
+ * @package Light Portal
+ * @link https://dragomano.ru/mods/light-portal
+ * @author Bugo <bugo@dragomano.ru>
+ * @copyright 2019-2020 Bugo
+ * @license https://spdx.org/licenses/GPL-3.0-or-later.html GPL-3.0-or-later
+ *
+ * @version 1.2
+ */
+
+if (!defined('SMF'))
+	die('Hacking attempt...');
+
+class PageArticle extends Article
+{
+	/**
+	 * Get active pages of the portal
+	 *
+	 * Получаем активные страницы портала
+	 *
+	 * @param int $start
+	 * @param int $limit
+	 * @return array
+	 */
+	public static function getData(int $start, int $limit): array
+	{
+		global $user_info, $smcFunc, $modSettings, $scripturl;
+
+		if (($pages = Helpers::cache()->get('articles_u' . $user_info['id'] . '_' . $start . '_' . $limit, LP_CACHE_TIME)) === null) {
+			$titles = Helpers::cache('all_titles', 'getAllTitles', '\Bugo\LightPortal\Subs', LP_CACHE_TIME, 'page');
+
+			$custom_columns = [];
+			$custom_tables  = [];
+			$custom_wheres  = [];
+
+			$custom_parameters = [
+				'type'         => 'page',
+				'status'       => \Bugo\LightPortal\Page::STATUS_ACTIVE,
+				'current_time' => time(),
+				'permissions'  => Helpers::getPermissions(),
+				'start'        => $start,
+				'limit'        => $limit
+			];
+
+			$custom_sorting = [
+				'comment_date DESC',
+				'p.created_at DESC',
+				'p.created_at',
+			];
+
+			Subs::runAddons('frontPages', array(&$custom_columns, &$custom_tables, &$custom_wheres, &$custom_parameters, &$custom_sorting));
+
+			$request = $smcFunc['db_query']('', '
+				SELECT
+					p.page_id, p.author_id, p.alias, p.content, p.description, p.type, p.status, p.num_views, p.num_comments, p.created_at, GREATEST(p.created_at, p.updated_at) AS date,
+					mem.real_name AS author_name, (
+						SELECT created_at
+						FROM {db_prefix}lp_comments
+						WHERE page_id = p.page_id
+						ORDER BY created_at DESC
+						LIMIT 1
+					) AS comment_date, (
+						SELECT author_id
+						FROM {db_prefix}lp_comments
+						WHERE created_at = comment_date
+						LIMIT 1
+					) AS comment_author_id, (
+						SELECT real_name
+						FROM {db_prefix}members
+						WHERE id_member = comment_author_id
+						LIMIT 1
+					) AS comment_author_name' . (!empty($custom_columns) ? ',
+					' . implode(', ', $custom_columns) : '') . '
+				FROM {db_prefix}lp_pages AS p
+					LEFT JOIN {db_prefix}members AS mem ON (p.author_id = mem.id_member)' . (!empty($custom_tables) ? '
+					' . implode("\n\t\t\t\t\t", $custom_tables) : '') . '
+				WHERE p.status = {int:status}
+					AND p.created_at <= {int:current_time}
+					AND p.permissions IN ({array_int:permissions})' . (!empty($custom_wheres) ? '
+					' . implode("\n\t\t\t\t\t", $custom_wheres) : '') . '
+				ORDER BY ' . (!empty($modSettings['lp_frontpage_order_by_num_replies']) ? 'IF (num_comments > 0, 0, 1), num_comments DESC, ' : '') . $custom_sorting[$modSettings['lp_frontpage_article_sorting'] ?? 0] . '
+				LIMIT {int:start}, {int:limit}',
+				$custom_parameters
+			);
+
+			$pages = [];
+			while ($row = $smcFunc['db_fetch_assoc']($request)) {
+				Helpers::parseContent($row['content'], $row['type']);
+
+				$image = null;
+				if (!empty($modSettings['lp_show_images_in_articles'])) {
+					$first_post_image = preg_match('/<img(.*)src(.*)=(.*)"(.*)"/U', $row['content'], $value);
+					$image = $first_post_image ? array_pop($value) : null;
+				}
+
+				if (!isset($pages[$row['page_id']])) {
+					$pages[$row['page_id']] = array(
+						'id'           => $row['page_id'],
+						'author_id'    => $author_id = empty($row['num_comments']) ? $row['author_id'] : $row['comment_author_id'],
+						'author_link'  => $scripturl . '?action=profile;u=' . $author_id,
+						'author_name'  => empty($row['num_comments']) ? $row['author_name'] : $row['comment_author_name'],
+						'teaser'       => Helpers::getTeaser($row['description'] ?: strip_tags($row['content'])),
+						'type'         => $row['type'],
+						'num_views'    => $row['num_views'],
+						'num_comments' => $row['num_comments'],
+						'date'         => empty($modSettings['lp_frontpage_article_sorting']) && !empty($row['comment_date']) ? $row['comment_date'] : $row['created_at'],
+						'is_new'       => $user_info['last_login'] < $row['date'] && $row['author_id'] != $user_info['id'],
+						'link'         => $scripturl . '?page=' . $row['alias'],
+						'image'        => $image,
+						'can_edit'     => $user_info['is_admin'] || (allowedTo('light_portal_manage_own_pages') && $row['author_id'] == $user_info['id'])
+					);
+				}
+
+				$pages[$row['page_id']]['title'] = $titles[$row['page_id']];
+
+				Subs::runAddons('frontPagesOutput', array(&$pages, $row));
+			}
+
+			$smcFunc['db_free_result']($request);
+			$smcFunc['lp_num_queries']++;
+
+			Helpers::cache()->put('articles_u' . $user_info['id'] . '_' . $start . '_' . $limit, $pages, LP_CACHE_TIME);
+		}
+
+		return $pages;
+	}
+
+	/**
+	 * Get count of active pages
+	 *
+	 * Получаем количество активных страниц
+	 *
+	 * @return int
+	 */
+	public static function getTotal(): int
+	{
+		global $user_info, $smcFunc;
+
+		if (($num_pages = Helpers::cache()->get('articles_u' . $user_info['id'] . '_total', LP_CACHE_TIME)) === null) {
+			$request = $smcFunc['db_query']('', '
+				SELECT COUNT(page_id)
+				FROM {db_prefix}lp_pages
+				WHERE status = {int:status}
+					AND created_at <= {int:current_time}
+					AND permissions IN ({array_int:permissions})',
+				array(
+					'status'       => \Bugo\LightPortal\Page::STATUS_ACTIVE,
+					'current_time' => time(),
+					'permissions'  => Helpers::getPermissions()
+				)
+			);
+
+			[$num_pages] = $smcFunc['db_fetch_row']($request);
+
+			$smcFunc['db_free_result']($request);
+			$smcFunc['lp_num_queries']++;
+
+			Helpers::cache()->put('articles_u' . $user_info['id'] . '_total', (int) $num_pages, LP_CACHE_TIME);
+		}
+
+		return (int) $num_pages;
+	}
+}
